*cmake, *ci clang++

+++ iow/io/socket/ad_initialize.hpp
@@ -14,7 +14,7 @@ struct ad_initialize
   {
     if ( opt.receive_buffer_size != 0 )
     {
-      boost::asio::socket_base::receive_buffer_size option(opt.receive_buffer_size);
+      boost::asio::socket_base::receive_buffer_size option( static_cast<int>(opt.receive_buffer_size) );
       boost::system::error_code ec;
       t.descriptor().set_option(option, ec);
       this->check(ec, "receive_buffer_size", opt.receive_buffer_size);
@@ -22,7 +22,7 @@ struct ad_initialize
 
     if ( opt.send_buffer_size != 0 )
     {
-      boost::asio::socket_base::send_buffer_size option(opt.send_buffer_size);
+      boost::asio::socket_base::send_buffer_size option( static_cast<int>(opt.send_buffer_size) );
       boost::system::error_code ec;
       t.descriptor().set_option(option, ec);
       this->check(ec, "send_buffer_size", opt.send_buffer_size);
